fix vc build
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@779 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/ViewItemRadar.cxx
@@ -338,7 +338,7 @@ void					ViewItemRadar::drawTank(float x, float y, float z, float minsize)
 
 	// Does not change with height.
 #ifdef WIN32
-	s = std::_cpp_max(atof(BZDB->get("tankRadius").c_str()), minsize);
+	s = std::_cpp_max((float)atof(BZDB->get("tankRadius").c_str()), minsize);
 #else
 	s = std::max(float(atof(BZDB->get("tankRadius").c_str())), minsize);
 #endif
